Issue #757: Test header interaction for all APIs; stop setting return address when response is not wanted.
