+++ services/connectivity/messaging/src/test/java/org/eclipse/ditto/services/connectivity/messaging/MessageMappingProcessorActorTest.java
@@ -13,10 +13,8 @@
 package org.eclipse.ditto.services.connectivity.messaging;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.ditto.services.connectivity.messaging.TestConstants.Authorization.AUTHORIZATION_CONTEXT;
 import static org.eclipse.ditto.services.connectivity.messaging.TestConstants.disableLogging;
 
-import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -26,12 +24,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
 
-import javax.annotation.Nullable;
-
-import org.assertj.core.api.AutoCloseableSoftAssertions;
 import org.eclipse.ditto.json.JsonFactory;
 import org.eclipse.ditto.json.JsonFieldSelector;
 import org.eclipse.ditto.json.JsonObject;
@@ -46,45 +39,30 @@ import org.eclipse.ditto.model.base.auth.AuthorizationModelFactory;
 import org.eclipse.ditto.model.base.auth.AuthorizationSubject;
 import org.eclipse.ditto.model.base.auth.DittoAuthorizationContextType;
 import org.eclipse.ditto.model.base.common.HttpStatusCode;
-import org.eclipse.ditto.model.base.common.ResponseType;
 import org.eclipse.ditto.model.base.headers.DittoHeaderDefinition;
 import org.eclipse.ditto.model.base.headers.DittoHeaders;
-import org.eclipse.ditto.model.base.headers.WithDittoHeaders;
-import org.eclipse.ditto.model.connectivity.Connection;
-import org.eclipse.ditto.model.connectivity.ConnectionId;
 import org.eclipse.ditto.model.connectivity.ConnectionSignalIdEnforcementFailedException;
 import org.eclipse.ditto.model.connectivity.ConnectivityModelFactory;
 import org.eclipse.ditto.model.connectivity.Enforcement;
 import org.eclipse.ditto.model.connectivity.EnforcementFactoryFactory;
 import org.eclipse.ditto.model.connectivity.EnforcementFilter;
 import org.eclipse.ditto.model.connectivity.EnforcementFilterFactory;
-import org.eclipse.ditto.model.connectivity.HeaderMapping;
-import org.eclipse.ditto.model.connectivity.MappingContext;
 import org.eclipse.ditto.model.connectivity.MessageMappingFailedException;
-import org.eclipse.ditto.model.connectivity.PayloadMapping;
-import org.eclipse.ditto.model.connectivity.PayloadMappingDefinition;
 import org.eclipse.ditto.model.connectivity.Target;
 import org.eclipse.ditto.model.connectivity.Topic;
-import org.eclipse.ditto.model.placeholders.Placeholder;
 import org.eclipse.ditto.model.placeholders.UnresolvedPlaceholderException;
 import org.eclipse.ditto.model.things.Thing;
-import org.eclipse.ditto.model.things.ThingId;
-import org.eclipse.ditto.protocoladapter.DittoProtocolAdapter;
 import org.eclipse.ditto.protocoladapter.JsonifiableAdaptable;
 import org.eclipse.ditto.protocoladapter.ProtocolFactory;
 import org.eclipse.ditto.protocoladapter.TopicPath;
-import org.eclipse.ditto.services.connectivity.mapping.ConnectivityCachingSignalEnrichmentProvider;
 import org.eclipse.ditto.services.connectivity.messaging.BaseClientActor.PublishMappedMessage;
 import org.eclipse.ditto.services.models.connectivity.ExternalMessage;
 import org.eclipse.ditto.services.models.connectivity.ExternalMessageFactory;
 import org.eclipse.ditto.services.models.connectivity.OutboundSignal;
 import org.eclipse.ditto.services.models.connectivity.OutboundSignalFactory;
-import org.eclipse.ditto.services.utils.akka.logging.DittoDiagnosticLoggingAdapter;
-import org.eclipse.ditto.services.utils.protocol.ProtocolAdapterProvider;
 import org.eclipse.ditto.signals.acks.base.Acknowledgement;
 import org.eclipse.ditto.signals.acks.base.Acknowledgements;
 import org.eclipse.ditto.signals.base.Signal;
-import org.eclipse.ditto.signals.commands.things.ThingErrorResponse;
 import org.eclipse.ditto.signals.commands.things.exceptions.ThingNotAccessibleException;
 import org.eclipse.ditto.signals.commands.things.modify.DeleteThingResponse;
 import org.eclipse.ditto.signals.commands.things.modify.ModifyAttribute;
@@ -92,73 +70,16 @@ import org.eclipse.ditto.signals.commands.things.modify.ModifyAttributeResponse;
 import org.eclipse.ditto.signals.commands.things.query.RetrieveThing;
 import org.eclipse.ditto.signals.commands.things.query.RetrieveThingResponse;
 import org.eclipse.ditto.signals.commands.thingsearch.subscription.CancelSubscription;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mockito;
-
-import com.typesafe.config.ConfigValueFactory;
 
 import akka.actor.ActorRef;
-import akka.actor.ActorSelection;
-import akka.actor.ActorSystem;
-import akka.actor.Props;
-import akka.pattern.Patterns;
 import akka.testkit.TestProbe;
 import akka.testkit.javadsl.TestKit;
 
 /**
  * Tests {@link MessageMappingProcessorActor}.
  */
-public final class MessageMappingProcessorActorTest {
-
-    private static final ThingId KNOWN_THING_ID = ThingId.of("my:thing");
-
-    private static final Connection CONNECTION = TestConstants.createConnectionWithAcknowledgements();
-    private static final ConnectionId CONNECTION_ID = CONNECTION.getId();
-
-    private static final DittoProtocolAdapter DITTO_PROTOCOL_ADAPTER = DittoProtocolAdapter.newInstance();
-    private static final String FAULTY_MAPPER = FaultyMessageMapper.ALIAS;
-    private static final String ADD_HEADER_MAPPER = AddHeaderMessageMapper.ALIAS;
-    private static final String DUPLICATING_MAPPER = DuplicatingMessageMapper.ALIAS;
-    private static final AuthorizationContext AUTHORIZATION_CONTEXT_WITH_DUPLICATES =
-            TestConstants.Authorization.withUnprefixedSubjects(AUTHORIZATION_CONTEXT);
-    private static final DittoHeaders headersWithReplyInformation = DittoHeaders.newBuilder()
-            .replyTarget(0)
-            .expectedResponseTypes(ResponseType.RESPONSE, ResponseType.ERROR, ResponseType.NACK)
-            .build();
-
-    private static final HeaderMapping CORRELATION_ID_AND_SOURCE_HEADER_MAPPING =
-            ConnectivityModelFactory.newHeaderMapping(JsonObject.newBuilder()
-                    .set("correlation-id", "{{ header:correlation-id }}")
-                    .set("source", "{{ request:subjectId }}")
-                    .build());
-
-    private static final HeaderMapping SOURCE_HEADER_MAPPING =
-            ConnectivityModelFactory.newHeaderMapping(JsonObject.newBuilder()
-                    .set("source", "{{ request:subjectId }}")
-                    .set("qos", "{{ header:qos }}")
-                    .build());
-
-    private ActorSystem actorSystem;
-    private ProtocolAdapterProvider protocolAdapterProvider;
-    private TestProbe connectionActorProbe;
-
-    @Before
-    public void setUp() {
-        actorSystem = ActorSystem.create("AkkaTestSystem", TestConstants.CONFIG);
-        protocolAdapterProvider = ProtocolAdapterProvider.load(TestConstants.PROTOCOL_CONFIG, actorSystem);
-        connectionActorProbe = TestProbe.apply("connectionActor", actorSystem);
-        MockProxyActor.create(actorSystem);
-    }
-
-    @After
-    public void tearDown() {
-        if (actorSystem != null) {
-            TestKit.shutdownActorSystem(actorSystem, scala.concurrent.duration.Duration.apply(5, TimeUnit.SECONDS),
-                    false);
-        }
-    }
+public final class MessageMappingProcessorActorTest extends AbstractMessageMappingProcessorActorTest {
 
     @Test
     public void testExternalMessageInDittoProtocolIsProcessedWithDefaultMapper() {
@@ -258,24 +179,6 @@ public final class MessageMappingProcessorActorTest {
             expectPublishedMappedMessage(publishMappedMessage, i, signal, targetWithEnrichment,
                     mapped -> assertThat(mapped.getAdaptable().getPayload().getExtra()).contains(extra));
         }};
-    }
-
-    @SafeVarargs
-    private static void expectPublishedMappedMessage(final PublishMappedMessage publishMappedMessage,
-            final int index,
-            final Signal<?> signal,
-            final Target target,
-            final Consumer<OutboundSignal.Mapped>... otherAssertionConsumers) {
-
-        final OutboundSignal.Mapped mapped =
-                publishMappedMessage.getOutboundSignal().getMappedOutboundSignals().get(index);
-        try (final AutoCloseableSoftAssertions softly = new AutoCloseableSoftAssertions()) {
-            softly.assertThat(mapped).satisfies(outboundSignal -> {
-                softly.assertThat(outboundSignal.getSource()).as("source is expected").isEqualTo(signal);
-                softly.assertThat(outboundSignal.getTargets()).as("targets are expected").containsExactly(target);
-            });
-        }
-        Arrays.asList(otherAssertionConsumers).forEach(con -> con.accept(mapped));
     }
 
     @Test
@@ -440,92 +343,6 @@ public final class MessageMappingProcessorActorTest {
         }};
     }
 
-    private void resetActorSystemWithCachingSignalEnrichmentProvider() {
-        TestKit.shutdownActorSystem(actorSystem);
-        actorSystem = ActorSystem.create("AkkaTestSystemWithCachingSignalEnrichmentProvider",
-                TestConstants.CONFIG
-                        .withValue("ditto.connectivity.signal-enrichment.provider",
-                                ConfigValueFactory.fromAnyRef(
-                                        ConnectivityCachingSignalEnrichmentProvider.class.getCanonicalName())
-                        )
-        );
-        MockProxyActor.create(actorSystem);
-    }
-
-    private void testExternalMessageInDittoProtocolIsProcessed(
-            @Nullable final EnforcementFilter<CharSequence> enforcement) {
-        testExternalMessageInDittoProtocolIsProcessed(enforcement, null);
-    }
-
-    private void testExternalMessageInDittoProtocolIsProcessed(
-            @Nullable final EnforcementFilter<CharSequence> enforcement, @Nullable final String mapping) {
-        testExternalMessageInDittoProtocolIsProcessed(enforcement, true, mapping, r -> {});
-    }
-
-    private void testExternalMessageInDittoProtocolIsProcessed(
-            @Nullable final EnforcementFilter<CharSequence> enforcement, final boolean expectSuccess,
-            @Nullable final String mapping, final Consumer<ThingErrorResponse> verifyErrorResponse) {
-
-        new TestKit(actorSystem) {{
-            final ActorRef messageMappingProcessorActor = createMessageMappingProcessorActor(this);
-            final ModifyAttribute modifyCommand = createModifyAttributeCommand();
-            final PayloadMapping mappings = ConnectivityModelFactory.newPayloadMapping(mapping);
-            final ExternalMessage externalMessage =
-                    ExternalMessageFactory.newExternalMessageBuilder(modifyCommand.getDittoHeaders())
-                            .withText(ProtocolFactory
-                                    .wrapAsJsonifiableAdaptable(DITTO_PROTOCOL_ADAPTER.toAdaptable(modifyCommand))
-                                    .toJsonString())
-                            .withAuthorizationContext(AUTHORIZATION_CONTEXT)
-                            .withEnforcement(enforcement)
-                            .withPayloadMapping(mappings)
-                            .withInternalHeaders(headersWithReplyInformation)
-                            .build();
-
-            TestProbe collectorProbe = TestProbe.apply("collector", actorSystem);
-            messageMappingProcessorActor.tell(externalMessage, collectorProbe.ref());
-
-            if (expectSuccess) {
-                final ModifyAttribute modifyAttribute = expectMsgClass(ModifyAttribute.class);
-                assertThat(modifyAttribute.getType()).isEqualTo(ModifyAttribute.TYPE);
-                assertThat(modifyAttribute.getDittoHeaders().getCorrelationId()).contains(
-                        modifyCommand.getDittoHeaders().getCorrelationId().orElse(null));
-                assertThat(modifyAttribute.getDittoHeaders().getAuthorizationContext())
-                        .isEqualTo(AUTHORIZATION_CONTEXT_WITH_DUPLICATES);
-                // thing ID is included in the header for error reporting
-                assertThat(modifyAttribute.getDittoHeaders())
-                        .extracting(headers -> headers.get(DittoHeaderDefinition.ENTITY_ID.getKey()))
-                        .isEqualTo(KNOWN_THING_ID.toString());
-                // internal headers added by consumer actors are appended
-                assertThat(modifyAttribute.getDittoHeaders()).containsEntry("ditto-reply-target", "0");
-
-                final String expectedMapperHeader = mapping == null ? "default" : mapping;
-                assertThat(modifyAttribute.getDittoHeaders().getInboundPayloadMapper()).contains(expectedMapperHeader);
-
-                if (ADD_HEADER_MAPPER.equals(mapping)) {
-                    assertThat(modifyAttribute.getDittoHeaders()).contains(AddHeaderMessageMapper.INBOUND_HEADER);
-                }
-
-                final ModifyAttributeResponse commandResponse =
-                        ModifyAttributeResponse.modified(KNOWN_THING_ID, modifyAttribute.getAttributePointer(),
-                                modifyAttribute.getDittoHeaders());
-
-                messageMappingProcessorActor.tell(commandResponse, getRef());
-                final OutboundSignal.Mapped responseMessage =
-                        expectMsgClass(PublishMappedMessage.class).getOutboundSignal().first();
-
-                if (ADD_HEADER_MAPPER.equals(mapping)) {
-                    final Map<String, String> headers = responseMessage.getExternalMessage().getHeaders();
-                    assertThat(headers).contains(AddHeaderMessageMapper.OUTBOUND_HEADER);
-                }
-            } else {
-                final OutboundSignal errorResponse = expectMsgClass(PublishMappedMessage.class).getOutboundSignal();
-                assertThat(errorResponse.getSource()).isInstanceOf(ThingErrorResponse.class);
-                final ThingErrorResponse response = (ThingErrorResponse) errorResponse.getSource();
-                verifyErrorResponse.accept(response);
-            }
-        }};
-    }
-
     @Test
     public void testReplacementOfPlaceholders() {
         final String correlationId = UUID.randomUUID().toString();
@@ -577,7 +394,7 @@ public final class MessageMappingProcessorActorTest {
                     "}";
             final ExternalMessage inboundMessage =
                     ExternalMessageFactory.newExternalMessageBuilder(Collections.emptyMap())
-                            .withInternalHeaders(headersWithReplyInformation)
+                            .withInternalHeaders(HEADERS_WITH_REPLY_INFORMATION)
                             .withText(messageContent)
                             .withAuthorizationContext(authorizationContext)
                             .build();
@@ -614,34 +431,6 @@ public final class MessageMappingProcessorActorTest {
         });
     }
 
-    private <T> void testMessageMappingWithoutCorrelationId(
-            final AuthorizationContext context,
-            final Class<T> expectedMessageClass,
-            final Consumer<T> verifyReceivedMessage) {
-
-        new TestKit(actorSystem) {{
-            final ActorRef messageMappingProcessorActor = createMessageMappingProcessorActor(this);
-            final Map<String, String> headers = new HashMap<>();
-            headers.put("content-type", "application/json");
-            final ModifyAttribute modifyCommand = ModifyAttribute.of(KNOWN_THING_ID, JsonPointer.of("foo"),
-                    JsonValue.of(42), DittoHeaders.empty());
-            final JsonifiableAdaptable adaptable = ProtocolFactory
-                    .wrapAsJsonifiableAdaptable(DITTO_PROTOCOL_ADAPTER.toAdaptable(modifyCommand));
-            final ExternalMessage externalMessage = ExternalMessageFactory.newExternalMessageBuilder(headers)
-                    .withTopicPath(adaptable.getTopicPath())
-                    .withText(adaptable.toJsonString())
-                    .withAuthorizationContext(context)
-                    .withHeaderMapping(SOURCE_HEADER_MAPPING)
-                    .build();
-
-            final TestProbe collectorProbe = TestProbe.apply("collector", actorSystem);
-            messageMappingProcessorActor.tell(externalMessage, collectorProbe.ref());
-
-            final T received = expectMsgClass(expectedMessageClass);
-            verifyReceivedMessage.accept(received);
-        }};
-    }
-
     @Test
     public void testTopicOnLiveTopicPathCombinationError() {
         final String correlationId = UUID.randomUUID().toString();
@@ -664,7 +453,7 @@ public final class MessageMappingProcessorActorTest {
                     "}";
             final ExternalMessage inboundMessage =
                     ExternalMessageFactory.newExternalMessageBuilder(Collections.emptyMap())
-                            .withInternalHeaders(headersWithReplyInformation)
+                            .withInternalHeaders(HEADERS_WITH_REPLY_INFORMATION)
                             .withText(messageContent)
                             .withAuthorizationContext(authorizationContext)
                             .build();
@@ -704,36 +493,6 @@ public final class MessageMappingProcessorActorTest {
                 });
     }
 
-    private <T> void testMessageMapping(final String correlationId,
-            final AuthorizationContext context,
-            final Class<T> expectedMessageClass,
-            final Consumer<T> verifyReceivedMessage) {
-
-        new TestKit(actorSystem) {{
-            final ActorRef messageMappingProcessorActor = createMessageMappingProcessorActor(this);
-            final Map<String, String> headers = new HashMap<>();
-            headers.put("correlation-id", correlationId);
-            headers.put("content-type", "application/json");
-            final ModifyAttribute modifyCommand = ModifyAttribute.of(KNOWN_THING_ID, JsonPointer.of("foo"),
-                    JsonValue.of(42), DittoHeaders.empty());
-            final JsonifiableAdaptable adaptable = ProtocolFactory
-                    .wrapAsJsonifiableAdaptable(DITTO_PROTOCOL_ADAPTER.toAdaptable(modifyCommand));
-            final ExternalMessage externalMessage = ExternalMessageFactory.newExternalMessageBuilder(headers)
-                    .withInternalHeaders(headersWithReplyInformation)
-                    .withTopicPath(adaptable.getTopicPath())
-                    .withText(adaptable.toJsonString())
-                    .withAuthorizationContext(context)
-                    .withHeaderMapping(CORRELATION_ID_AND_SOURCE_HEADER_MAPPING)
-                    .build();
-
-            final TestProbe collectorProbe = TestProbe.apply("collector", actorSystem);
-            messageMappingProcessorActor.tell(externalMessage, collectorProbe.ref());
-
-            final T received = expectMsgClass(expectedMessageClass);
-            verifyReceivedMessage.accept(received);
-        }};
-    }
-
     @Test
     public void testCommandResponseIsProcessed() {
         new TestKit(actorSystem) {{
@@ -742,7 +501,7 @@ public final class MessageMappingProcessorActorTest {
             final String correlationId = UUID.randomUUID().toString();
             final ModifyAttributeResponse commandResponse =
                     ModifyAttributeResponse.modified(KNOWN_THING_ID, JsonPointer.of("foo"),
-                            headersWithReplyInformation.toBuilder().correlationId(correlationId).build());
+                            HEADERS_WITH_REPLY_INFORMATION.toBuilder().correlationId(correlationId).build());
 
             messageMappingProcessorActor.tell(commandResponse, getRef());
 
@@ -763,7 +522,7 @@ public final class MessageMappingProcessorActorTest {
             final String correlationId = UUID.randomUUID().toString();
             final ThingNotAccessibleException thingNotAccessibleException =
                     ThingNotAccessibleException.newBuilder(KNOWN_THING_ID)
-                            .dittoHeaders(headersWithReplyInformation.toBuilder()
+                            .dittoHeaders(HEADERS_WITH_REPLY_INFORMATION.toBuilder()
                                     .correlationId(correlationId)
                                     .putHeader(DittoHeaderDefinition.ENTITY_ID.getKey(), KNOWN_THING_ID)
                                     .build())
@@ -923,85 +682,4 @@ public final class MessageMappingProcessorActorTest {
             assertThat(received.getSubscriptionId()).isEqualTo(searchCommand.getSubscriptionId());
         }};
     }
-
-    private ActorRef createMessageMappingProcessorActor(final TestKit kit) {
-        final Props props =
-                MessageMappingProcessorActor.props(kit.getRef(), kit.getRef(), getMessageMappingProcessor(),
-                        CONNECTION, connectionActorProbe.ref(), 99);
-        return actorSystem.actorOf(props);
-    }
-
-    private MessageMappingProcessor getMessageMappingProcessor() {
-        final Map<String, MappingContext> mappingDefinitions = new HashMap<>();
-        mappingDefinitions.put(FAULTY_MAPPER, FaultyMessageMapper.CONTEXT);
-        mappingDefinitions.put(ADD_HEADER_MAPPER, AddHeaderMessageMapper.CONTEXT);
-        mappingDefinitions.put(DUPLICATING_MAPPER, DuplicatingMessageMapper.CONTEXT);
-        final PayloadMappingDefinition payloadMappingDefinition =
-                ConnectivityModelFactory.newPayloadMappingDefinition(mappingDefinitions);
-        final DittoDiagnosticLoggingAdapter logger = Mockito.mock(DittoDiagnosticLoggingAdapter.class);
-        Mockito.when(logger.withCorrelationId(Mockito.any(DittoHeaders.class)))
-                .thenReturn(logger);
-        Mockito.when(logger.withCorrelationId(Mockito.any(CharSequence.class)))
-                .thenReturn(logger);
-        Mockito.when(logger.withCorrelationId(Mockito.any(WithDittoHeaders.class)))
-                .thenReturn(logger);
-        return MessageMappingProcessor.of(CONNECTION_ID, payloadMappingDefinition, actorSystem,
-                TestConstants.CONNECTIVITY_CONFIG,
-                protocolAdapterProvider, logger);
-    }
-
-    private void setUpProxyActor(final ActorRef recipient) {
-        final ActorSelection actorSelection = actorSystem.actorSelection("/user/connectivityRoot" +
-                "/connectivityProxyActor");
-        Patterns.ask(actorSelection, recipient, Duration.ofSeconds(10L))
-                .toCompletableFuture()
-                .join();
-    }
-
-    private ExternalMessage toExternalMessage(final Signal<?> signal) {
-        final AuthorizationContext context =
-                AuthorizationModelFactory.newAuthContext(
-                        DittoAuthorizationContextType.UNSPECIFIED,
-                        AuthorizationModelFactory.newAuthSubject("ditto:ditto"));
-        final JsonifiableAdaptable adaptable = ProtocolFactory
-                .wrapAsJsonifiableAdaptable(DITTO_PROTOCOL_ADAPTER.toAdaptable(signal));
-        return ExternalMessageFactory.newExternalMessageBuilder(Map.of())
-                .withTopicPath(adaptable.getTopicPath())
-                .withText(adaptable.toJsonString())
-                .withAuthorizationContext(context)
-                .build();
-    }
-
-    private static ModifyAttribute createModifyAttributeCommand() {
-        final Map<String, String> headers = new HashMap<>();
-        final String correlationId = UUID.randomUUID().toString();
-        headers.put("correlation-id", correlationId);
-        headers.put("content-type", "application/json");
-        return ModifyAttribute.of(KNOWN_THING_ID, JsonPointer.of("foo"), JsonValue.of(42), DittoHeaders.of(headers));
-    }
-
-    private static final class TestPlaceholder implements Placeholder<String> {
-
-        @Override
-        public String getPrefix() {
-            return "test";
-        }
-
-        @Override
-        public List<String> getSupportedNames() {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public boolean supports(final String name) {
-            return true;
-        }
-
-        @Override
-        public Optional<String> resolve(final String placeholderSource, final String name) {
-            return Optional.of(placeholderSource);
-        }
-
-    }
-
 }
