+++ services/gateway/endpoints/src/test/java/org/eclipse/ditto/services/gateway/endpoints/actors/HttpRequestActorTest.java
@@ -16,60 +16,28 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.nio.charset.Charset;
 import java.util.Collections;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
-import java.util.stream.Collectors;
 
-import javax.annotation.Nullable;
-
-import org.eclipse.ditto.json.JsonObject;
 import org.eclipse.ditto.json.JsonPointer;
 import org.eclipse.ditto.json.JsonValue;
 import org.eclipse.ditto.model.base.acks.AcknowledgementRequest;
 import org.eclipse.ditto.model.base.acks.DittoAcknowledgementLabel;
-import org.eclipse.ditto.model.base.auth.AuthorizationContext;
-import org.eclipse.ditto.model.base.auth.AuthorizationModelFactory;
-import org.eclipse.ditto.model.base.auth.AuthorizationSubject;
-import org.eclipse.ditto.model.base.auth.DittoAuthorizationContextType;
-import org.eclipse.ditto.model.base.common.HttpStatusCode;
-import org.eclipse.ditto.model.base.headers.DittoHeaderDefinition;
 import org.eclipse.ditto.model.base.headers.DittoHeaders;
-import org.eclipse.ditto.model.messages.Message;
-import org.eclipse.ditto.model.messages.MessageDirection;
-import org.eclipse.ditto.model.messages.MessageHeaderDefinition;
-import org.eclipse.ditto.model.messages.MessageHeaders;
-import org.eclipse.ditto.model.policies.SubjectId;
-import org.eclipse.ditto.model.policies.SubjectIssuer;
 import org.eclipse.ditto.model.things.ThingId;
-import org.eclipse.ditto.protocoladapter.HeaderTranslator;
-import org.eclipse.ditto.services.gateway.endpoints.routes.whoami.DefaultUserInformation;
-import org.eclipse.ditto.services.gateway.endpoints.routes.whoami.UserInformation;
 import org.eclipse.ditto.services.gateway.endpoints.routes.whoami.Whoami;
-import org.eclipse.ditto.services.gateway.util.config.DittoGatewayConfig;
-import org.eclipse.ditto.services.gateway.util.config.GatewayConfig;
-import org.eclipse.ditto.services.utils.config.DefaultScopedConfig;
-import org.eclipse.ditto.signals.commands.messages.SendThingMessage;
 import org.eclipse.ditto.signals.commands.messages.SendThingMessageResponse;
-import org.eclipse.ditto.signals.commands.things.modify.ModifyAttribute;
 import org.eclipse.ditto.signals.commands.things.modify.ModifyAttributeResponse;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.typesafe.config.ConfigFactory;
-
 import akka.actor.ActorRef;
-import akka.actor.ActorSystem;
 import akka.http.javadsl.model.ContentTypes;
 import akka.http.javadsl.model.HttpEntities;
-import akka.http.javadsl.model.HttpHeader;
 import akka.http.javadsl.model.HttpRequest;
 import akka.http.javadsl.model.HttpResponse;
 import akka.http.javadsl.model.ResponseEntity;
 import akka.http.javadsl.model.StatusCode;
 import akka.http.javadsl.model.StatusCodes;
-import akka.http.javadsl.model.headers.RawHeader;
 import akka.testkit.TestProbe;
 import akka.testkit.javadsl.TestKit;
 import akka.util.ByteString;
@@ -77,28 +45,7 @@ import akka.util.ByteString;
 /**
  * Unit test for {@link HttpRequestActor}.
  */
-public final class HttpRequestActorTest {
-
-    private static final HeaderTranslator HEADER_TRANSLATOR = HeaderTranslator.of(
-            DittoHeaderDefinition.values(),
-            MessageHeaderDefinition.values());
-
-    private static ActorSystem system;
-    private static GatewayConfig gatewayConfig;
-
-    @BeforeClass
-    public static void beforeClass() {
-        system = ActorSystem.create();
-        final DefaultScopedConfig dittoConfig = DefaultScopedConfig.dittoScoped(ConfigFactory.load("test.conf"));
-        gatewayConfig = DittoGatewayConfig.of(dittoConfig);
-    }
-
-    @AfterClass
-    public static void afterClass() {
-        if (null != system) {
-            TestKit.shutdownActorSystem(system);
-        }
-    }
+public final class HttpRequestActorTest extends AbstractHttpRequestActorTest {
 
     @Test
     public void handlesWhoamiCommand() throws ExecutionException, InterruptedException {
@@ -132,63 +79,6 @@ public final class HttpRequestActorTest {
 
             final StatusCode expectedHttpStatusCode = StatusCodes.NO_CONTENT;
             final boolean expectHttpResponseHeaders = true;
-
-            testThingModifyCommand(thingId, attributeName, attributePointer, dittoHeaders, expectedHeaders,
-                    probeResponse, expectedHttpStatusCode, expectHttpResponseHeaders);
-        }};
-    }
-
-    @Test
-    public void handlesThingModifyCommandNoResponseRequired() throws ExecutionException, InterruptedException {
-        new TestKit(system) {{
-            final ThingId thingId = ThingId.generateRandom();
-            final String attributeName = "foo";
-            final JsonPointer attributePointer = JsonPointer.of(attributeName);
-
-            final DittoHeaders dittoHeaders = createAuthorizedHeaders()
-                    .toBuilder()
-                    .responseRequired(false)
-                    .build();
-
-            final DittoHeaders expectedHeaders = dittoHeaders.toBuilder()
-                    .acknowledgementRequests(Collections.emptySet())
-                    .build();
-
-            // no response is sent to the HttpRequestActor as none was required and "twin-persisted" ack label was not
-            //  requested:
-            final ModifyAttributeResponse probeResponse = null;
-
-            final StatusCode expectedHttpStatusCode = StatusCodes.ACCEPTED;
-            final boolean expectHttpResponseHeaders = false;
-
-            testThingModifyCommand(thingId, attributeName, attributePointer, dittoHeaders, expectedHeaders,
-                    probeResponse, expectedHttpStatusCode, expectHttpResponseHeaders);
-        }};
-    }
-
-    @Test
-    public void handlesThingModifyCommandNoResponseRequiredAndTwinPersistedAckRequest()
-            throws ExecutionException, InterruptedException {
-        new TestKit(system) {{
-            final ThingId thingId = ThingId.generateRandom();
-            final String attributeName = "foo";
-            final JsonPointer attributePointer = JsonPointer.of(attributeName);
-
-            final DittoHeaders dittoHeaders = createAuthorizedHeaders()
-                    .toBuilder()
-                    .responseRequired(false)
-                    .acknowledgementRequest(AcknowledgementRequest.of(DittoAcknowledgementLabel.TWIN_PERSISTED))
-                    .build();
-            // headers shall be same:
-            final DittoHeaders expectedHeaders = dittoHeaders.toBuilder()
-                    .build();
-
-            // response is sent to the HttpRequestActor as "twin-persisted" ack label was requested:
-            final ModifyAttributeResponse probeResponse =
-                    ModifyAttributeResponse.modified(thingId, attributePointer, expectedHeaders);
-
-            final StatusCode expectedHttpStatusCode = StatusCodes.ACCEPTED;
-            final boolean expectHttpResponseHeaders = false;
 
             testThingModifyCommand(thingId, attributeName, attributePointer, dittoHeaders, expectedHeaders,
                     probeResponse, expectedHttpStatusCode, expectHttpResponseHeaders);
@@ -274,46 +164,6 @@ public final class HttpRequestActorTest {
             testThingModifyCommand(thingId, attributeName, attributePointer, dittoHeaders, expectedHeaders,
                     probeResponse, expectedHttpStatusCode, expectHttpResponseHeaders);
         }};
-    }
-
-    private void testThingModifyCommand(final ThingId thingId,
-            final String attributeName,
-            final JsonPointer attributePointer,
-            final DittoHeaders dittoHeaders,
-            final DittoHeaders expectedHeaders,
-            @Nullable final ModifyAttributeResponse probeResponse,
-            final StatusCode expectedHttpStatusCode,
-            final boolean expectHttpResponseHeaders) throws InterruptedException, ExecutionException {
-
-        final TestProbe proxyActorProbe = TestProbe.apply(system);
-
-        final ModifyAttribute modifyAttribute = ModifyAttribute.of(
-                thingId, attributePointer, JsonValue.of("bar"), dittoHeaders);
-
-        final HttpRequest request =
-                HttpRequest.PUT("/api/2/things/" + thingId.toString() + "/attributes/" + attributeName);
-        final CompletableFuture<HttpResponse> responseFuture = new CompletableFuture<>();
-
-        final ActorRef underTest = createHttpRequestActor(proxyActorProbe.ref(), request, responseFuture);
-        underTest.tell(modifyAttribute, ActorRef.noSender());
-
-        final ModifyAttribute expectedModifyAttribute = modifyAttribute.setDittoHeaders(expectedHeaders);
-        proxyActorProbe.expectMsg(expectedModifyAttribute);
-
-        if (null != probeResponse) {
-            proxyActorProbe.reply(probeResponse);
-        }
-
-        HttpResponse expectedHttpResponse = HttpResponse.create().withStatus(expectedHttpStatusCode);
-        if (expectHttpResponseHeaders) {
-            final List<HttpHeader> expectedHttpResponseHeaders = HEADER_TRANSLATOR.toExternalHeaders(expectedHeaders)
-                    .entrySet()
-                    .stream()
-                    .map(e -> RawHeader.create(e.getKey(), e.getValue()))
-                    .collect(Collectors.toList());
-            expectedHttpResponse = expectedHttpResponse.withHeaders(expectedHttpResponseHeaders);
-        }
-        assertThat(responseFuture.get()).isEqualTo(expectedHttpResponse);
     }
 
     @Test
@@ -407,115 +257,4 @@ public final class HttpRequestActorTest {
                     probeResponse, expectedHttpStatusCode, expectHttpResponseHeaders, expectedHttpResponseEntity);
         }};
     }
-
-    private void testMessageCommand(final ThingId thingId,
-            final String messageSubject,
-            final DittoHeaders dittoHeaders,
-            final DittoHeaders expectedHeaders,
-            @Nullable final SendThingMessageResponse probeResponse,
-            final StatusCode expectedHttpStatusCode,
-            final boolean expectHttpResponseHeaders,
-            @Nullable final ResponseEntity expectedHttpResponseEntity) throws InterruptedException, ExecutionException {
-
-        final TestProbe proxyActorProbe = TestProbe.apply(system);
-
-        final Message<?> message = Message.newBuilder(
-                MessageHeaders.newBuilder(MessageDirection.TO, thingId, messageSubject)
-                        .contentType("application/json")
-                        .build()
-        )
-                .payload(JsonValue.of("ping"))
-                .build();
-        final SendThingMessage<?> sendThingMessage = SendThingMessage.of(thingId, message, dittoHeaders);
-
-        final HttpRequest request =
-                HttpRequest.PUT("/api/2/things/" + thingId.toString() + "/inbox/messages/" + messageSubject);
-        final CompletableFuture<HttpResponse> responseFuture = new CompletableFuture<>();
-
-        final ActorRef underTest = createHttpRequestActor(proxyActorProbe.ref(), request, responseFuture);
-        underTest.tell(sendThingMessage, ActorRef.noSender());
-
-        final SendThingMessage<?> expectedSendThingMessage = sendThingMessage.setDittoHeaders(expectedHeaders);
-        proxyActorProbe.expectMsg(expectedSendThingMessage);
-
-        if (null != probeResponse) {
-            proxyActorProbe.reply(probeResponse);
-        }
-
-        HttpResponse expectedHttpResponse = HttpResponse.create().withStatus(expectedHttpStatusCode);
-        if (expectHttpResponseHeaders) {
-            final List<HttpHeader> expectedHttpResponseHeaders = HEADER_TRANSLATOR.toExternalHeaders(expectedHeaders)
-                    .entrySet()
-                    .stream()
-                    .map(e -> RawHeader.create(e.getKey(), e.getValue()))
-                    .collect(Collectors.toList());
-            expectedHttpResponse = expectedHttpResponse.withHeaders(expectedHttpResponseHeaders);
-        }
-        if (null != expectedHttpResponseEntity) {
-            expectedHttpResponse = expectedHttpResponse.withEntity(expectedHttpResponseEntity);
-        }
-        assertThat(responseFuture.get()).isEqualTo(expectedHttpResponse);
-    }
-
-    private SendThingMessageResponse<?> buildSendThingMessageResponse(final ThingId thingId,
-            final String messageSubject,
-            final DittoHeaders expectedHeaders,
-            final String contentType,
-            final JsonValue responsePayload) {
-        final Message<?> responseMessage = Message.newBuilder(
-                MessageHeaders.newBuilder(MessageDirection.TO, thingId, messageSubject)
-                        .statusCode(HttpStatusCode.IMUSED)
-                        .contentType(contentType)
-                        .build()
-        )
-                .payload(responsePayload)
-                .build();
-        return SendThingMessageResponse.of(thingId, responseMessage, HttpStatusCode.IMUSED, expectedHeaders);
-    }
-
-    private ActorRef createHttpRequestActor(final ActorRef proxyActorRef, final HttpRequest request,
-            final CompletableFuture<HttpResponse> response) {
-        return system.actorOf(HttpRequestActor.props(
-                proxyActorRef,
-                HEADER_TRANSLATOR,
-                request,
-                response,
-                gatewayConfig.getHttpConfig(),
-                gatewayConfig.getCommandConfig()
-        ));
-    }
-
-    private DittoHeaders createAuthorizedHeaders() {
-        final AuthorizationContext context = AuthorizationContext.newInstance(DittoAuthorizationContextType.JWT,
-                AuthorizationSubject.newInstance(SubjectId.newInstance(SubjectIssuer.GOOGLE, "any-google-user")),
-                AuthorizationSubject.newInstance(
-                        SubjectId.newInstance(SubjectIssuer.INTEGRATION, "any-integration-subject")));
-        return DittoHeaders.newBuilder()
-                .randomCorrelationId()
-                .authorizationContext(context)
-                .build();
-    }
-
-    private HttpResponse createExpectedWhoamiResponse(final Whoami whoami) {
-        final AuthorizationContext authContext =
-                getAuthContextWithPrefixedSubjectsFromHeaders(whoami.getDittoHeaders());
-        final UserInformation userInformation = DefaultUserInformation.fromAuthorizationContext(authContext);
-        final List<HttpHeader> expectedHeaders = HEADER_TRANSLATOR.toExternalHeaders(whoami.getDittoHeaders())
-                .entrySet()
-                .stream()
-                .map(e -> RawHeader.create(e.getKey(), e.getValue()))
-                .collect(Collectors.toList());
-        return HttpResponse.create().withStatus(StatusCodes.OK)
-                .withEntity(ContentTypes.APPLICATION_JSON, ByteString.fromString(userInformation.toJsonString()))
-                .withHeaders(expectedHeaders);
-    }
-
-    private static AuthorizationContext getAuthContextWithPrefixedSubjectsFromHeaders(final DittoHeaders headers) {
-        final String authContextString = headers.get(DittoHeaderDefinition.AUTHORIZATION_CONTEXT.getKey());
-        final JsonObject authContextJson = authContextString == null ?
-                JsonObject.empty() :
-                JsonObject.of(authContextString);
-        return AuthorizationModelFactory.newAuthContext(authContextJson);
-    }
-
 }
