+++ services/gateway/endpoints/src/test/resources/test.conf
@@ -3,6 +3,7 @@ akka {
   loggers = ["akka.event.slf4j.Slf4jLogger"]
   loglevel = "WARNING"
   logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
+  log-dead-letters = 0
 
   # for log messages during the actor system is starting up and shutting down:
   stdout-loglevel = "WARNING"
